+++ src/parser/ast/comment_block.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+pub struct CommentBlock {
+    pub(crate) name: Option<String>,
+    pub(crate) desc: Option<String>,
+}