+++ src/parser/ast/mod.rs
@@ -27,3 +27,4 @@ pub(crate) mod reference;
 pub(crate) mod constant;
 pub(crate) mod entity;
 pub(crate) mod arith_expr;
+pub(crate) mod comment_block;