+++ src/parser/ast/model.rs
@@ -1,3 +1,4 @@
+use crate::parser::ast::comment_block::CommentBlock;
 use crate::parser::ast::decorator::Decorator;
 use crate::parser::ast::field::Field;
 use crate::parser::ast::identifier::Identifier;
@@ -8,6 +9,7 @@ pub struct Model {
     pub(crate) id: usize,
     pub(crate) source_id: usize,
     pub(crate) identifier: Identifier,
+    pub(crate) comment_block: Option<CommentBlock>,
     pub(crate) fields: Vec<Field>,
     pub(crate) decorators: Vec<Decorator>,
     pub(crate) span: Span,
@@ -18,9 +20,9 @@ pub struct Model {
 }
 
 impl Model {
-    pub(crate) fn new(id: usize, source_id: usize, identifier: Identifier, fields: Vec<Field>, decorators: Vec<Decorator>, span: Span) -> Self {
+    pub(crate) fn new(id: usize, source_id: usize, identifier: Identifier, comment_block: Option<CommentBlock>, fields: Vec<Field>, decorators: Vec<Decorator>, span: Span) -> Self {
         Self {
-            id, source_id, identifier, fields, decorators, span, resolved: false,
+            id, source_id, identifier, comment_block, fields, decorators, span, resolved: false,
             scalar_field_enum: vec![], scalar_field_and_cached_property_enum: vec![],
             direct_relation_enum: vec![],
         }